Support for 'where' clause in filter_subtable
New 'where' syntax is more powerful than old polymorphic
syntax with 'type' and 'type_key'. We now can support
any expression that sequel supports.

The old syntax is deprecated and remove. I decided to
go for simplicity and less code over convinience of a
syntactic sugar. Existing Shrinkfiles will have to be
rewritten, as this is incompatible change. Version
number updated to reflect that.